SnPM Batch Script
This script will run the Matlab toolbox SnPM on your second level
analyses to find cluster thresholds for your analysis. Must have SnPM
installed.

You do not need to make any changes if you do not want to. This file
assumes that you are finding thresholds on second-level analyses and
that the model name is the name of the folder where the first-level
contrasts are (based on the structure of
SUBJECT -> MODEL -> contrasts.con

You will be prompted to first choose a directory to output to

Next you will be prompted to choose a directory that has your
second-level analyses in it. The script will select all SPM.mat files
in any folder within that directory.

Third you will be asked to select the thresholds you would like to
check at. You may choose more than one. If you want to select different
thresholds, you must write them in the thresholds variable as a cell of
strings

Finally you will choose what models/contrasts you actually want to run.
If you want to run every one just select all,